+++ test/Client/AzureManaged.Tests/DurableTaskSchedulerClientExtensionsTests.cs
@@ -89,7 +89,6 @@ public class DurableTaskSchedulerClientExtensionsTests
         workerOptions.EndpointAddress.Should().Be("http://localhost");
         workerOptions.TaskHubName.Should().Be(ValidTaskHub);
         workerOptions.Credential.Should().BeNull();
-        workerOptions.ResourceId.Should().Be("https://durabletask.io");
         workerOptions.AllowInsecureCredentials.Should().BeTrue();
     }
 
@@ -192,7 +191,6 @@ public class DurableTaskSchedulerClientExtensionsTests
         options.EndpointAddress.Should().Be(ValidEndpoint); // The https:// prefix is added by CreateChannel, not in the extension method
         options.TaskHubName.Should().Be(ValidTaskHub);
         options.Credential.Should().BeOfType<DefaultAzureCredential>();
-        options.ResourceId.Should().Be("https://durabletask.io");
         options.AllowInsecureCredentials.Should().BeFalse();
     }